gcp-cloud-function-samples/LoadTriggeredBucketFileIntoBigQuery/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-api-python-client
pandas
fsspec
gcsfs
pyarrow